+++ web-frontend/modules/core/components/notifications/Notifications.vue
@@ -20,6 +20,7 @@
         :state="undoRedoState"
       ></UndoRedoNotification>
       <CopyingNotification v-if="copying"></CopyingNotification>
+      <PastingNotification v-if="pasting"></PastingNotification>
       <RestoreNotification
         v-for="notification in restoreNotifications"
         :key="notification.id"
@@ -37,6 +38,7 @@ import ConnectingNotification from '@baserow/modules/core/components/notificatio
 import FailedConnectingNotification from '@baserow/modules/core/components/notifications/FailedConnectingNotification'
 import RestoreNotification from '@baserow/modules/core/components/notifications/RestoreNotification'
 import CopyingNotification from '@baserow/modules/core/components/notifications/CopyingNotification'
+import PastingNotification from '@baserow/modules/core/components/notifications/PastingNotification'
 import AuthorizationErrorNotification from '@baserow/modules/core/components/notifications/AuthorizationErrorNotification'
 import UndoRedoNotification from '@baserow/modules/core/components/notifications/UndoRedoNotification'
 import { UNDO_REDO_STATES } from '@baserow/modules/core/utils/undoRedoConstants'
@@ -49,6 +51,7 @@ export default {
     ConnectingNotification,
     FailedConnectingNotification,
     CopyingNotification,
+    PastingNotification,
     AuthorizationErrorNotification,
     UndoRedoNotification,
   },
@@ -67,6 +70,7 @@ export default {
       connecting: (state) => state.notification.connecting,
       failedConnecting: (state) => state.notification.failedConnecting,
       copying: (state) => state.notification.copying,
+      pasting: (state) => state.notification.pasting,
       notifications: (state) => state.notification.items,
       undoRedoState: (state) => state.notification.undoRedoState,
     }),