+++ web-frontend/modules/database/mixins/gridField.js
@@ -1,6 +1,7 @@
 import { isElement } from '@baserow/modules/core/utils/dom'
 import { copyToClipboard } from '@baserow/modules/database/utils/clipboard'
 import baseField from '@baserow/modules/database/mixins/baseField'
+import Papa from 'papaparse'
 
 /**
  * A mixin that can be used by a field grid component. It introduces the props that
@@ -140,7 +141,8 @@ export default {
           const value = this.$registry
             .get('field', this.field.type)
             .prepareValueForCopy(this.field, rawValue)
-          copyToClipboard(value)
+          const tsv = Papa.unparse([[value]], { delimiter: '\t' })
+          copyToClipboard(tsv)
         }
 
         // Removes the value if the backspace/delete key is pressed.
@@ -162,23 +164,40 @@ export default {
       document.body.addEventListener('keydown', this.$el.keyDownEvent)
 
       // Updates the value of the field when a user pastes something in the field.
-      this.$el.pasteEvent = (event) => {
+      this.$el.pasteEvent = async (event) => {
         if (!this.canPaste(event)) {
           return
         }
 
-        const value = this.$registry
-          .get('field', this.field.type)
-          .prepareValueForPaste(this.field, event.clipboardData)
-        const oldValue = this.value
-        if (
-          value !== undefined &&
-          value !== oldValue &&
-          !this.readOnly &&
-          !this.field._.type.isReadOnly
-        ) {
-          this.$emit('update', value, oldValue)
-        }
+        try {
+          // Multiple values in TSV format can be provided, so we need to properly
+          // parse it using Papa.
+          const parsed = await Papa.parsePromise(
+            event.clipboardData.getData('text'),
+            { delimiter: '\t' }
+          )
+          const data = parsed.data
+          // A grid field cell can only handle one single value. We try to extract
+          // that from the clipboard and update the cell, otherwise we emit the
+          // paste event up.
+          if (data.length === 1 && data[0].length === 1) {
+            const value = this.$registry
+              .get('field', this.field.type)
+              .prepareValueForPaste(this.field, data[0][0])
+            const oldValue = this.value
+
+            if (
+              value !== undefined &&
+              value !== oldValue &&
+              !this.readOnly &&
+              !this.field._.type.isReadOnly
+            ) {
+              this.$emit('update', value, oldValue)
+            }
+          } else {
+            this.$emit('paste', data)
+          }
+        } catch (e) {}
       }
       document.addEventListener('paste', this.$el.pasteEvent)